Initial commit, genresSaga and Slice added. Genres removed from external sagas and slices.
