+++ src/features/movieList/movieListSlice.js
@@ -5,7 +5,6 @@ const movieListSlice = createSlice({
   initialState: {
     status: "initial",
     movies: null,
-    genres: null,
     query: null,
     page: null,
     totalResults: 0,
@@ -24,13 +23,6 @@ const movieListSlice = createSlice({
     fetchMoviesError: (state) => {
       state.status = "error";
     },
-    fetchGenres: () => { },
-    fetchGenresSuccess: (state, { payload: genres }) => {
-      state.genres = genres;
-    },
-    fetchGenresError: (state) => {
-      state.genres = null;
-    },
     setQuery: (state, { payload: query }) => {
       state.query = query;
       state.status = "loading";
@@ -45,9 +37,6 @@ export const {
   fetchMovies,
   fetchMoviesSuccess,
   fetchMoviesError,
-  fetchGenres,
-  fetchGenresError,
-  fetchGenresSuccess,
   setQuery,
   setPage,
 } = movieListSlice.actions;
@@ -56,8 +45,6 @@ const selectMovieListState = (state) => state.movieList;
 
 export const selectStatus = (state) => selectMovieListState(state).status;
 export const selectMovies = (state) => selectMovieListState(state).movies;
-export const selectGenres = (state) => selectMovieListState(state).genres;
-export const selectIsGenres = (state) => selectMovieListState(state).isGenres;
 export const selectPage = (state) => selectMovieListState(state).page;
 export const selectQuery = (state) => selectMovieListState(state).query;
 export const selectTotalResults = (state) => selectMovieListState(state).totalResults;
